feat(dto): Add DTOs for all entities and refactor author model
Introduces a complete set of Data Transfer Objects (DTOs) for all domain entities to define clear API contracts and separates request/response models.

- Creates `Request` DTOs (e.g., `UserRegisterRequestDTO`, `BookCreateRequestDTO`) with validation annotations.
- Creates `Response` DTOs (e.g., `UserResponseDTO`, `BookResponseDTO`) to securely expose data.
- Refactors the `Book` entity to use a dedicated `Author` entity instead of a primitive String, normalizing the database schema.
- Establishes a clear separation between the domain model and the API layer.

+++ src/main/java/com/eduardoxduardo/vlibrary/model/entities/Review.java
@@ -23,7 +23,7 @@ public class Review implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private int rating;
+    private Double rating;
     private String comment;
 
     @Column(name = "review_date")
@@ -33,7 +33,7 @@ public class Review implements Serializable {
     @JoinColumn(name = "user_book_id", nullable = false, unique = true)
     private UserBook userBook;
 
-    public Review(int rating, String comment, UserBook userBook) {
+    public Review(double rating, String comment, UserBook userBook) {
         this.rating = rating;
         this.comment = comment;
         this.reviewDate = LocalDate.now();
